[#34]refactor: 렌더시 조인 이외 렌더 방식으로 변경
