+++ src/components/recommend/Recommend.js
@@ -1,15 +1,15 @@
 const Component = require('../../lib/Component');
 
-class Aside extends Component {
+class Recommend extends Component {
   constructor({ parent }) {
     super(parent);
   }
 
   render() {
     return `<div class="recommend-aside">
-                ${this.components.map((component) => component.render()).join('')}
+                ${this.renderToString()}
             </div>`;
   }
 }
 
-module.exports = Aside;
+module.exports = Recommend;
